Implement automatic notification tap handling in plugin
- Move notification intent handling from MainActivity to PntaFlutterPlugin
- Add NewIntentListener interface implementation to plugin
- Simplify MainActivity to basic FlutterActivity extension
- Automatically process notification tap payloads without requiring custom MainActivity code
- Add proper activity binding lifecycle management
